MainModule/requirements.txt
panda3d==1.10.14
paho-mqtt==2.1.1
numpy==2.0.0
